Started making basic functionality

+++ NorthwindEntitiesLib/Product.cs
@@ -20,23 +20,23 @@ namespace NorthwindApi.AutoGen
 
         [Key]
         [Column("ProductID")]
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         [Column(TypeName = "nvarchar (40)")]
         public string ProductName { get; set; } = null!;
         [Column("SupplierID", TypeName = "int")]
-        public long? SupplierId { get; set; }
+        public int? SupplierId { get; set; }
         [Column("CategoryID", TypeName = "int")]
-        public long? CategoryId { get; set; }
+        public int? CategoryId { get; set; }
         [Column(TypeName = "nvarchar (20)")]
         public string? QuantityPerUnit { get; set; }
         [Column(TypeName = "money")]
         public byte[]? UnitPrice { get; set; }
         [Column(TypeName = "smallint")]
-        public long? UnitsInStock { get; set; }
+        public int? UnitsInStock { get; set; }
         [Column(TypeName = "smallint")]
-        public long? UnitsOnOrder { get; set; }
+        public int? UnitsOnOrder { get; set; }
         [Column(TypeName = "smallint")]
-        public long? ReorderLevel { get; set; }
+        public int? ReorderLevel { get; set; }
         [Column(TypeName = "bit")]
         public byte[] Discontinued { get; set; } = null!;
